Use `rt_uint_ptr' for `op_third/op_size'.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@48482 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/ipc/ewb/eif_in.c
@@ -162,7 +162,7 @@ rt_public void send_breakpoint (BODY_INDEX real_body_id, long int offset, EIF_BO
 	rqst.rq_type = BREAK_ON;
 	rqst.rq_opaque.op_first = (int) real_body_index;
 	rqst.rq_opaque.op_second = (int) real_body_id;
-	rqst.rq_opaque.op_third = (long) offset;
+	rqst.rq_opaque.op_third = offset;
 
 #ifdef EIF_WIN32
 	if (-1 == send_packet (sp, &rqst))
